Move NavButton into separate include/, src/ files
